Graphics.Matplotlib.Internal

Description

Internal representations of the Matplotlib data. These are not API-stable and may change. You can easily extend the provided bindings without relying on the internals exposed here but they are provided just in case.

Documentation

mapLinear :: (Double -> b) -> Double -> Double -> Double -> [b]

A handy miscellaneous function to linearly map over a range of numbers in a given number of steps

Basics

data Matplotlib

The wrapper type for a matplotlib computation.

Constructors

Matplotlib

Fields

mpCommands :: Seq MplotCommand

Resolved computations that have been transformed to commands

mpPendingOption :: Maybe ([Option] -> MplotCommand)

A pending computation that is affected by applied options

mpRest :: Seq MplotCommand

Computations that follow the one that is pending

data MplotCommand

A maplotlib command, right now we have a very shallow embedding essentially dealing in strings containing python code as well as the ability to load data. The loaded data should be a json object.

Constructors

LoadData ByteString

forall x.MplotImage x => LoadImage x

Exec

Fields

es :: String

data Option

Throughout the API we need to accept options in order to expose matplotlib's many configuration options.

Constructors

K String String
results in a=b

P String
just inserts the option verbatim as an argument at the end of the function

toPy :: MplotCommand -> String

Convert an MplotCommand to python code, doesn't do much right now

resolvePending :: Matplotlib -> Matplotlib

Resolve the pending command with no options provided.

withMplot :: Matplotlib -> ([String] -> IO a) -> IO a

The io action is given a list of python commands to execute (note that these are commands in the sense of lines of python code; each inidivudal line may not be parseable on its own)

mplotString :: String -> Matplotlib

Create a plot that executes the string as python code

mp :: Matplotlib

Create an empty plot. This the beginning of most plotting commands.

readData :: ToJSON a => a -> Matplotlib

Load the given data into the python "data" array

readImage :: MplotImage i => i -> Matplotlib

Load the given image into python "img" variable

(%) :: Matplotlib -> Matplotlib -> Matplotlib infixl 5

Combine two matplotlib commands

(#) :: MplotValue val => Matplotlib -> val -> Matplotlib infixl 6

Add Python code to the last matplotlib command

data S

A string to be rendered in python as a string. In other words it is rendered as str.

Constructors

S String

data R

A string to be rendered in python as a raw string. In other words it is rendered as rstr.

Constructors

R String

data L

A string to be rendered in python as a raw literal/code. In other words it is inserted directly as is into the code.

Constructors

L String

class MplotValue val where

Values which can be combined together to form a matplotlib command. These specify how values are rendered in Python code.

Minimal complete definition

toPython

Methods

toPython :: val -> String

Render a value inline in Python code

toPythonOpt :: val -> String

Render a value as an optional parameter in Python code

class MplotImage a where

The class of Haskell images or references to imagese which can be transferred to matplotlib.

Methods

saveHaskellImage :: a -> FilePath -> IO String

loadPythonImage :: a -> String -> FilePath -> String

Options

optFn :: ([Option] -> String) -> Matplotlib -> Matplotlib

Add an option to the last matplotlib command. Commands can have only one option! optFn :: Matplotlib -> Matplotlib

options :: Matplotlib -> Matplotlib

Merge two commands with options between

(##) :: MplotValue val => Matplotlib -> val -> Matplotlib infixl 6

A combinator like # that also inserts an option

renderOptions :: [Option] -> [Char]

An internal helper to convert a list of options to the python code that applies those options in a call.

optionFn :: ([Option] -> [Option]) -> Matplotlib -> Matplotlib

An internal helper that modifies the options of a plot.

option :: Matplotlib -> [Option] -> Matplotlib

Apply a list of options to a plot resolving any pending options.

(@@) :: Matplotlib -> [Option] -> Matplotlib infixl 6

A combinator for option that applies a list of options to a plot

bindDefault :: Matplotlib -> [Option] -> Matplotlib

Bind a list of default options to a plot. Positional options are kept in order and default that way as well. Keyword arguments are also handled

bindDefaultFn :: [Option] -> [Option] -> [Option]

Merge two sets of options

Python operations

python :: Foldable t => t String -> IO (Either String String)

Run python given a code string.

pyBackend :: [Char] -> [Char]

pyIncludes :: String -> [[Char]]

The standard python includes of every plot

escapeSlashes :: [Char] -> [Char]

These will be Python strings and slashes would cause unwanted control characters.

pyReadData :: [Char] -> [[Char]]

The python command that reads external data into the python data array

pyReadImage :: [Char] -> [[Char]]

The python command that reads an image into the img variable

pyDetach :: [[Char]]

Detach python so we don't block (TODO This isn't working reliably)

pyOnscreen :: [[Char]]

Python code to show a plot

pyFigure :: [Char] -> [[Char]]

Python code that saves a figure

pySVG :: [[Char]]

Python code that returns SVG for a figure

o1 :: MplotValue val => val -> Option

Create a positional option

o2 :: MplotValue a => String -> a -> Option

Create a keyword option

str :: String -> S

Create a string that will be rendered as a python string

raw :: String -> R

Create a string that will be rendered as a raw python string

lit :: String -> L

Create a literal that will inserted into the python code directly

updateAxes :: Matplotlib

Update axes. Should be called any time the state is changed.

updateFigure :: Matplotlib

Update the figure and the axes. Should be called any time the state is changed.

minimum2 :: (Ord (t a), Ord a, Foldable t1, Foldable t) => t1 (t a) -> a

Smallest element of a list of lists

maximum2 :: (Ord (t a), Ord a, Foldable t1, Foldable t) => t1 (t a) -> a

Largest element of a list of lists